move validateData to helpers dir
