+++ helpers/index.js
@@ -1,2 +1,3 @@
 export * from "./setupFolder.js";
 export * from "./cropAndResize.js";
+export * from "./validateData.js";
