+++ validators/index.js
@@ -1,4 +1,3 @@
 export * from "./subscriptionSchema.js";
-export * from "./validate.js";
 export * from "./userSchema.js";
 export * from "./contactSchema.js";
